Move PollWithRetries to core module
