+++ java/clients/src/main/java/sleeper/clients/teardown/TerminateEMRClusters.java
@@ -25,8 +25,8 @@ import com.amazonaws.services.s3.AmazonS3ClientBuilder;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import sleeper.clients.util.PollWithRetries;
 import sleeper.configuration.properties.InstanceProperties;
+import sleeper.core.util.PollWithRetries;
 
 import java.io.IOException;
 import java.util.List;
