+++ java/core/src/main/java/sleeper/core/util/PollWithRetries.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.clients.util;
+package sleeper.core.util;
 
 import com.google.common.math.LongMath;
