+++ java/system-test/src/main/java/sleeper/systemtest/bulkimport/WaitForEMRClusters.java
@@ -26,8 +26,8 @@ import com.amazonaws.services.s3.AmazonS3ClientBuilder;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import sleeper.clients.util.PollWithRetries;
 import sleeper.configuration.properties.InstanceProperties;
+import sleeper.core.util.PollWithRetries;
 import sleeper.systemtest.SystemTestProperties;
 
 import java.io.IOException;
