+++ java/system-test/src/main/java/sleeper/systemtest/compaction/WaitForCompactionJobs.java
@@ -22,10 +22,10 @@ import com.amazonaws.services.s3.AmazonS3ClientBuilder;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import sleeper.clients.util.PollWithRetries;
 import sleeper.compaction.job.CompactionJobStatusStore;
 import sleeper.compaction.job.status.CompactionJobStatus;
 import sleeper.compaction.status.store.job.CompactionJobStatusStoreFactory;
+import sleeper.core.util.PollWithRetries;
 import sleeper.systemtest.SystemTestProperties;
 
 import java.io.IOException;
