+++ java/system-test/src/main/java/sleeper/systemtest/ingest/WaitForGenerateData.java
@@ -24,7 +24,7 @@ import com.amazonaws.services.ecs.model.Task;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import sleeper.clients.util.PollWithRetries;
+import sleeper.core.util.PollWithRetries;
 import sleeper.systemtest.ingest.json.TaskStatusJson;
 import sleeper.systemtest.ingest.json.TasksJson;
