+++ java/system-test/src/test/java/sleeper/systemtest/util/WaitForQueueEstimateTest.java
@@ -19,7 +19,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
-import sleeper.clients.util.PollWithRetries;
 import sleeper.compaction.job.CompactionJob;
 import sleeper.compaction.job.CompactionJobStatusStore;
 import sleeper.compaction.job.CompactionJobTestDataHelper;
@@ -27,6 +26,7 @@ import sleeper.compaction.testutils.CompactionJobStatusStoreInMemory;
 import sleeper.configuration.properties.InstanceProperties;
 import sleeper.configuration.properties.InstanceProperty;
 import sleeper.core.record.process.RecordsProcessedSummary;
+import sleeper.core.util.PollWithRetries;
 import sleeper.job.common.QueueMessageCount;
 
 import java.time.Duration;
